% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_coldiff.R
\name{gen_coldiff}
\alias{gen_coldiff}
\title{Generate column difference indicator}
\usage{
gen_coldiff(data, cols, engine = "identical", ...)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}

\item{engine}{A character vector indicating the function used to test for equality. One of either "identical" for \href{https://rdrr.io/r/base/identical.html}{\code{base::identical()}} or "all.equal" for \href{https://rdrr.io/r/base/all.equal.html}{\code{base::all.equal()}}.}

\item{...}{Arguments passed onto the function defined in \code{engine}. For example, the \code{tolerance} argument can be passed onto \code{base::all.equal()} to ignore numeric differences smaller than \code{tolerance}.}
}
\value{
An integer vector of value integer 1 (\code{1L}) for different columns or an integer 0 (\code{0L}) for identical (or near-equal) columns.
}
\description{
This function returns a binary indicator of whether any selected columns are different.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = 1:3,
  y = as.double(1:3),
  z = 1.0001:3.0001
)
gen_coldiff(a, c(x, y))
b <- a \%>\%
  mutate(
    q = gen_coldiff(., c(x, y)),
    r = gen_coldiff(., c(x, y), engine = "all.equal"),
    s = gen_coldiff(., c(x, z), engine = "all.equal", tolerance = 1e-4),
    t = gen_coldiff(., c(x, z), engine = "all.equal", tolerance = 1e-5)
  )
b
}
