% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_na_listcol.R
\name{gen_na_listcol}
\alias{gen_na_listcol}
\title{Replace NULLs in a list-column with NAs}
\usage{
gen_na_listcol(.x)
}
\arguments{
\item{.x}{A list or list-column to modify.}
}
\value{
A list with all \code{NULL} values replaced with \code{NA}.
}
\description{
This function takes a list and replaces all \code{NULL} values with \code{NA}. It is useful for working with list-columns in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <-
  mtcars \%>\%
  select(cyl, vs, am) \%>\%
  slice(1:6) \%>\%
  as_tibble() \%>\%
  mutate(listcol = list(NULL, "b", "c", "d", "e", "f"))
glimpse(a)
b <-
  a \%>\%
  mutate(across(starts_with("listcol"), gen_na_listcol))
glimpse(b)
}
