% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_percent.R
\name{gen_percent}
\alias{gen_percent}
\title{Generate column percent}
\usage{
gen_percent(data, col, by, prop = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{col}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> A single column with which to calculate a column percent.}

\item{by}{An optional character vector of columns to group by.}

\item{prop}{If \code{TRUE}, percent will be shown as a proportion between 0-1 rather than a percent between 0-100. Default is \code{FALSE}.}
}
\value{
A double vector totaling 100 within \code{col}. If grouping columns are specified with \code{by}, the percent for each unique category of grouping columns will total 100 within \code{col}. If \code{prop} is specified, a double vector totaling 1 within \code{col} (or totaling 1 within unique categories of grouping columns specified with \code{by}).
}
\description{
This function calculates a column percent. The \code{by} argument calculates column percents within unique categories of grouping columns. The \code{prop} argument calculates a proportion rather than a percent.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- as_tibble(mtcars)
gen_percent(a, gear)
b <-
  a \%>\%
  select(gear, cyl, carb) \%>\%
  arrange(gear, cyl, carb) \%>\%
  mutate(
    pct1 = gen_percent(., gear),
    pct2 = gen_percent(., gear, by = "cyl"),
    pct3 = gen_percent(., gear, by = c("cyl", "carb")),
    prop1 = gen_percent(., gear, prop = TRUE)
  )
b
}
