% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowall.R
\name{gen_rowall}
\alias{gen_rowall}
\title{Generate rowwise match of all values}
\usage{
gen_rowall(data, cols, values)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}

\item{values}{A list of values to match.}
}
\value{
A binary integer vector indicating whether all supplied values were matched with an integer 1 (\code{1L}), otherwise it returns an integer 0 (\code{0L}).
}
\description{
This function performs a rowwise match of all supplied values across columns in a data frame. If all of the row values equal one of the supplied values, this function returns an integer 1 (\code{1L}) for that row, otherwise it returns an integer 0 (\code{0L}).
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = 1:3,
  y = rep(NA, 3),
  z = letters[1:3],
  aa = rep(FALSE, 3)
)
val <- list(1, NA, "a", FALSE)
val2 <- list(3, NA, "c", FALSE)
gen_rowall(a, values = val)
b <- a \%>\%
  mutate(
    q = gen_rowall(., values = val),
    r = gen_rowall(., values = val2)
  )
b
}
