% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowcount.R
\name{gen_rowcount}
\alias{gen_rowcount}
\title{Rowwise count of the number of columns matching a set of values}
\usage{
gen_rowcount(data, cols, values)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}

\item{values}{A list of values to match.}
}
\value{
An integer vector with the number of matched values.
}
\description{
This function performs a rowwise count of columns in a data frame that match a set of supplied values.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = 1:3,
  y = rep(NA, 3),
  z = letters[1:3],
  aa = rep(FALSE, 3)
)
val <- list(1, NA, "a", FALSE)
gen_rowcount(a, values = val)
gen_rowcount(a, everything(), values = val)
gen_rowcount(a, starts_with(letters[25:26]), values = val)
b <- a \%>\% mutate(q = gen_rowcount(., values = val))
b
}
