% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowfirst.R
\name{gen_rowfirst}
\alias{gen_rowfirst}
\title{Generate rowwise first nonmissing value}
\usage{
gen_rowfirst(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
A vector of the rowwise first nonmissing value. The vector's type will be of common type to all rowwise nonmissing values.
}
\description{
This function returns the rowwise first nonmissing value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rowfirst(a)
gen_rowfirst(a, all_of(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowfirst(.))
b
c <-
  a \%>\%
  mutate(w = c("a", TRUE, NA), .before = "x") \%>\%
  mutate(q = gen_rowfirst(.))
c # note that q is of type <chr>
}
