% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowlast.R
\name{gen_rowlast}
\alias{gen_rowlast}
\title{Generate rowwise last nonmissing value}
\usage{
gen_rowlast(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
A vector of the rowwise last nonmissing value. The vector's type will be of common type to all rowwise nonmissing values.
}
\description{
This function returns the rowwise last nonmissing value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rowlast(a)
gen_rowlast(a, all_of(letters[24:25]))
b <- a \%>\% mutate(q = gen_rowlast(.))
b
c <-
  a \%>\%
  mutate(aa = c("a", TRUE, NA), .after = "z") \%>\%
  mutate(q = gen_rowlast(.))
c # note that q is of type <chr>
}
