% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowmax.R
\name{gen_rowmax}
\alias{gen_rowmax}
\title{Generate rowwise maximum value}
\usage{
gen_rowmax(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
A vector of the rowwise maximum value.
}
\description{
This function returns the rowwise maximum value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rowmax(a)
gen_rowmax(a, everything())
gen_rowmax(a, starts_with(letters[24:25]))
b <- a \%>\% mutate(q = gen_rowmax(.))
b
}
