% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowmean.R
\name{gen_rowmean}
\alias{gen_rowmean}
\title{Generate rowwise mean}
\usage{
gen_rowmean(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
A double vector of the rowwise arithmetic mean value. Missing values are ignored.
}
\description{
This function returns the rowwise arithmetic mean value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rowmean(a)
gen_rowmean(a, everything())
gen_rowmean(a, all_of(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowmean(.))
b
}
