% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowmedian.R
\name{gen_rowmedian}
\alias{gen_rowmedian}
\title{Generate rowwise median}
\usage{
gen_rowmedian(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
A double vector of the rowwise median value. Missing values are ignored.
}
\description{
This function returns the rowwise median value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(2, 3, 2),
  z = c(4, NA, 5)
)
gen_rowmedian(a)
gen_rowmedian(a, everything())
gen_rowmedian(a, all_of(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowmedian(.))
b
}
