% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowmin.R
\name{gen_rowmin}
\alias{gen_rowmin}
\title{Generate rowwise minimum value}
\usage{
gen_rowmin(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
A vector of the rowwise minimum value.
}
\description{
This function returns the rowwise minimum value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rowmin(a)
gen_rowmin(a, everything())
gen_rowmin(a, starts_with(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowmin(.))
b
}
