% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowmiss.R
\name{gen_rowmiss}
\alias{gen_rowmiss}
\title{Generate rowwise count of missing values}
\usage{
gen_rowmiss(data, cols)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}
}
\value{
An integer vector of the rowwise count of missing values.
}
\description{
This function returns the rowwise count of missing values in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rowmiss(a)
gen_rowmiss(a, all_of(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowmiss(.))
b
}
