% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rownth.R
\name{gen_rownth}
\alias{gen_rownth}
\title{Generate rowwise nth nonmissing value}
\usage{
gen_rownth(data, cols, n)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across.}

\item{n}{An integer vector of length 1 that specifies the position of the rowwise nth nonmissing value to search for. A negative integer will index from the end.}
}
\value{
A vector of the rowwise nth nonmissing value. The vector's type will be of common type to all rowwise nonmissing values.
}
\description{
This function returns the rowwise nth nonmissing value in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5)
)
gen_rownth(a, n = 1)
gen_rownth(a, n = 2)
gen_rownth(a, all_of(letters[25:26]), n = 1)
b <- a \%>\% mutate(q = gen_rownth(., n = 1), r = gen_rownth(., n = 2))
b
c <-
  a \%>\%
  mutate(w = c("a", TRUE, NA), .before = "x") \%>\%
  mutate(q = gen_rownth(., n = 1), r = gen_rownth(., n = 2))
c # note that q and r are of type <chr>
}
