% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowsd.R
\name{gen_rowsd}
\alias{gen_rowsd}
\title{Generate rowwise standard deviation}
\usage{
gen_rowsd(data, cols, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to search across. Non-numeric columns are ignored.}

\item{na.rm}{Logical vector of length one passed onto \href{https://rdrr.io/r/stats/sd.html}{\code{stats::sd()}}.}
}
\value{
A double vector of the rowwise standard deviation.
}
\description{
This function returns the rowwise standard deviation in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.

If \code{na.rm} is \code{TRUE} (the default), missing values are removed. If \code{na.rm} is \code{FALSE}, the presence of an \code{NA} or \code{NaN} in any row will return \code{NA} for that row. As per the documentation for \href{https://rdrr.io/r/stats/sd.html}{\code{stats::sd()}}, the standard deviation of a length one vector is \code{NA}.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5),
  aa = c("a", "b", "c"),
  bb = rep(NA, 3),
  cc = rep(NaN, 3)
)
gen_rowsd(a)
gen_rowsd(a, all_of(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowsd(.), r = gen_rowsd(., na.rm = FALSE))
b
}
