% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rowsum.R
\name{gen_rowsum}
\alias{gen_rowsum}
\title{Generate rowwise sum}
\usage{
gen_rowsum(data, cols, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{cols}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> Columns to sum. Non-numeric columns are ignored.}

\item{na.rm}{Logical vector of length one passed onto \href{https://rdrr.io/r/base/sum.html}{\code{base::sum()}}.}
}
\value{
A vector of the rowwise sum. The vector's type will be of common type to all rowwise numeric values.
}
\description{
This function returns the rowwise sum in a data frame.
}
\details{
Parallelization is supported via \href{https://purrr.tidyverse.org/reference/in_parallel.html}{\code{purrr::in_parallel()}}.

If \code{na.rm} is \code{TRUE} (the default), missing values are removed. If \code{na.rm} is \code{FALSE}, the presence of an \code{NA} or \code{NaN} in any row will return \code{NA} or \code{NaN} for that row.
}
\examples{
library(dplyr, warn.conflicts = FALSE)
a <- tibble(
  x = c(1, NA, 2),
  y = c(NA, 3, NA),
  z = c(4, NA, 5),
  aa = c("a", "b", "c"),
  bb = rep(NA, 3),
  cc = rep(NaN, 3)
)
gen_rowsum(a)
gen_rowsum(a, all_of(letters[25:26]))
b <- a \%>\% mutate(q = gen_rowsum(.), r = gen_rowsum(., na.rm = FALSE))
b
}
