% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutagen-package.R
\docType{package}
\name{mutagen-package}
\alias{mutagen}
\alias{mutagen-package}
\title{mutagen: Extensions to dplyr's mutate}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extensions to dplyr's mutate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gvelasq/mutagen}
  \item \url{https://gvelasq.github.io/mutagen/}
  \item Report bugs at \url{https://github.com/gvelasq/mutagen/issues}
}

}
\author{
\strong{Maintainer}: Gustavo Velásquez \email{gvelasq2github@gmail.com}

}
\keyword{internal}
